refactor: rename word2vec2 to word2vec
