+++ crates/w2v/src/lib.rs
@@ -1,1 +1,2 @@
-pub mod word2vec2;
+
+pub mod word2vec;
