+++ crates/w2v/src/word2vec.rs
